refactor: extract execution, dry-run, and state-helpers from orchestrator index
Splits orchestrator/index.ts (612 lines) into focused modules:
- execution.ts: runMainLoop + buildPRBody (339 lines)
- dry-run.ts: runDryRun (112 lines)
- state-helpers.ts: initState + getIssueState (35 lines)
- index.ts: entry point + parseFlags + re-exports (167 lines)

Breaks circular dependency via state-helpers.ts. Updates source
guard tests to cover new module locations. All 155 tests pass.
